feat(frontend): Some refactoring
Added datetimes to posts, add post sorting

IssueID #
