+++ src/components/bloglist.component.tsx
@@ -3,7 +3,7 @@ import { FC, ReactElement } from 'react';
 import { useQuery } from 'react-query';
 import { Link } from 'react-router-dom';
 
-import { IBlogPost } from './interfaces/interfaces';
+import { IBlogPost } from '../interfaces/interfaces';
 
 const BlogList: FC = (): ReactElement => {
   const { isLoading, isError, data, error } = useQuery(['blogs'], () => fetch('/blogs').then(res => res.json()));
@@ -16,6 +16,8 @@ const BlogList: FC = (): ReactElement => {
     return <h1>Error: {error.message}</h1>;
   }
 
+  data.sort((a: IBlogPost, b: IBlogPost) => (new Date(a.postdate) >= new Date(b.postdate) ? 1 : -1));
+  console.log('sorted', data);
   return (
     <div>
       {data.map((post: IBlogPost) => (
@@ -23,7 +25,9 @@ const BlogList: FC = (): ReactElement => {
           <Link to={`/blogs/${post.id}`}>
             <h1>{post.title}</h1>
           </Link>
-          <div>{post.author}</div>
+          <div>
+            by {post.author}, {post.postdate}
+          </div>
         </div>
       ))}
     </div>
